requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
weasyprint>=60.0
markdown>=3.5
pdf2image>=1.16.3
Pillow>=10.0.0
